Add Python and environment files to .gitignore
The .gitignore file tells Git which files and folders to ignore when committing code to the repository. This helps keep the repository clean, lightweight, and free of unnecessary files.

For this project, the .gitignore includes :
--> __pycache__/ and *.py[cod] – automatically created Python cache and compiled files.
--> venv/ – virtual environment folder with installed packages (not needed in repo).
--> .vscode/ – VSCode workspace settings.
--> .env – environment or secret configuration files.

By ignoring these files, we ensure the repository contains only the important project code and resources, making it easier to share, maintain, and collaborate.